[Refactor] 더미 데이터 분리 및 CircleShape 라운딩 적용
